Initial mock-up (for study) of IPV4 checksum layer feature extension.
Please see the IPV4 element in ethernetIP.dfdl.xsd, and the definitions
in the new ipv4checksumLayer.dfdl.xsd file.

Shows how IPV4 layer transforms can compute the checksum, and how the
schema has to evolve to incorporate the checksum when parsing and unparsing.

This feature is Daffodil ticket DAFFODIL-2221