Version 1.5.0 Remove excess namespaces. Change namespace URI.
Namespace URI now follows conventions.

Eliminate the bb, ipa namespaces.
Retain the one for the layer, but the rest
are not needed.

Principle here is that one schema should usually mean one namespace to understand
for all its formats, types, and groups to
be defined in.
We make an exception for layer definitions.

Update README.md with usage instructions.

https://github.com/DFDLSchemas/ethernetIP/issues/21